+++ userspace/lib/usbip_network.h
@@ -12,7 +12,7 @@ void usbip_setup_port_number(const char *arg);
 
 int usbip_net_recv(SOCKET sockfd, void *buff, size_t bufflen);
 int usbip_net_send(SOCKET sockfd, void *buff, size_t bufflen);
-int usbip_net_send_op_common(SOCKET sockfd, uint32_t code, uint32_t status);
+int usbip_net_send_op_common(SOCKET sockfd, uint16_t code, uint32_t status);
 int usbip_net_recv_op_common(SOCKET sockfd, uint16_t *code, int *pstatus);
 int usbip_net_set_reuseaddr(SOCKET sockfd);
 int usbip_net_set_nodelay(SOCKET sockfd);
